Decorators: fix issue with conditional skips and None
Callable conditional were introduced in f402b6d81, but it also
introduced a bug.  In an effort to support both non-conditionals
(plain "@skip") and conditionals ("@skipIf" and "@skipUnless")
a parameter called condition was introduced, and checked for being
None.

The problem is that None should evaluate to False, and thus cause
the skipIf to *not* skip, such as in:

   @avocado.skipIf(None,
                   'skipIf with None condition, should not happen')

And conversely, skipUnless to skip.  The bugfix simply makes the
non-conditional skip have a True condition instead.

Also added are tests that verify the behavior of skipIf/skipUnless
with both callables, non-callables and None, asserting that conditions
where tests should not be skipped are respected.

Signed-off-by: Cleber Rosa <crosa@redhat.com>