Sysinfo: support configuration files installed system wide
Commit 768455c3 removed the existing configuration files, and with
that, the patching that happened on the RPM SPEC file changing the
sysinfo defaults to the system wide /etc based locations.

Without a configuration file to tweak, we're left wither with patching
code, or making the code more dynamic and respecting the system wide
configuration paths no matter if it's an RPM based installation or
not.  This chooses the later.

Fixes: https://github.com/avocado-framework/avocado/issues/4208
Signed-off-by: Cleber Rosa <crosa@redhat.com>