nrunner: re-enable persistence of a executable tests' stdout/stderr
This re-enables the handling of stdout/stderr with the Exec based
runners.  It basically sends the two log messages, one for stdout and
one for stderr, as defined in the current documentation.

The processing of the final message is still given stdout and stderr
because some runners may judge the success/failure based on the
content on those buffers.

Signed-off-by: Cleber Rosa <crosa@redhat.com>