lint: spliting normal checks from the Python 3 port checks
According to the pylint community: py3k (Python 3 port mode) warnings
are intentionally left as disabled because they describe a whole other
type of checking. You should use pylint in py3k mode or not and it is
not intended to lint code and check it for Python3 compatibility at the
same time. Even that we are not fully supporting --py3k yet, lets split
this for a better organization.  Also, because of that, we are not using
pylintrc here, due to limitations in its syntax. Maybe in the future we
can use multiple pylintrc files.

Signed-off-by: Beraldo Leal <bleal@redhat.com>