Nrunner: preserve the output generated by Python unittests
Currently, no output generated by the python-unittest kind is
preserved as part of the test results.  While it's debatable if the
output is an "stdout", this change at least preserves the output
which are very important when things go wrong and tests fail.

Further work should formalize a more generic mechanism and naming
of output produced by different test types when there's no actual
stdout and/or stderr.

Signed-off-by: Cleber Rosa <crosa@redhat.com>