Pre release testplan: do not shutdown the virtual machine
The instruction to shutdown the virtual machine is a leftover from the
times we'd run a test with the libvirt runner, and expect the machine
to be powered up.

The next step on the testplan actually instructs to run a test on the
virtual machine, so this is even more conflicting.

Signed-off-by: Cleber Rosa <crosa@redhat.com>