Proper nrunner message saving
The saving of the nrunner messages was too strict. The stdout and stderr
messages could be only bytes and the log messages could be only strings
without a new line at the end. This commit makes the saving more
flexible. Every message is bytes and runner can send encoding if the
message should be decoded.

Signed-off-by: Jan Richter <jarichte@redhat.com>