Instrumented nrunner error handling
When the exception is raised on the side of the instrumented runner not
inside the test itself, the runner will crash and don't send any message
to the avocado and such an event is not logged into the error log. This
commit handles such event and logs more information about the occurred
exception.

Signed-off-by: Jan Richter <jarichte@redhat.com>