Task identifier as TestID
Some of our plugins expecting that when task is a test that the
identifier is `avocado.core.test_id.TestID`, but this is not enforced an
even the default identifier is not `TestID` but it is a string. This
commit adds the ability to convert `Task.identifier` to `TestID` when
it's needed.

Signed-off-by: Jan Richter <jarichte@redhat.com>