Logging handler: turn "core.show" from a list into a set
The goal of the "core.show" configuration is to keep a unique set
of the logging facilities to be shown.  We've been using a list,
but a set is really a better data structure to keep track of those.

To ease the migration process, especially with custom jobs, let's
still allow lists, but convent them to a set very early in the
application command line parsing.

Signed-off-by: Cleber Rosa <crosa@redhat.com>