Job: incorporate support for store logging streams
This is currently a "run" plugin feature, but it's really within the
scope of a job.

By moving the feature to the job, we not only put it in a better
place, we also eliminate a technical debt with regards to the use of
default values at get() time, and also remove the need for that manual
value in Job API based tests.

Signed-off-by: Cleber Rosa <crosa@redhat.com>