plugins.list: move TERM decorators to proper location
We should apply TERM colors only when displaying data for a better
cohesion. If we need to get the matrix without the colors (let's say
that I would like to save a clean json output), it is not possible
today, without some hacks. Yes, this is a preparation for future commits
where I will introduce the json output.

Signed-off-by: Beraldo Leal <bleal@redhat.com>