Allow test references to be set in a configuraiton file
When integrating Avocado into existing test suites, it can make sense
to provide the list of test references in a configuration file.

This allows users to simply do:

 $ avocado --config=.avocado.config run

This is a first step towards enabling Avocado to also read
configuration on de-facto project files, such as setup.cfg and
pyproject.toml.  When that is done, users would be able to
simply do:

 $ avocado run

If the project contains the valid configuration files.

Signed-off-by: Cleber Rosa <crosa@redhat.com>