Empty reference in resolver
Some test might be able to be resolved without a reference. Config
files or other parameters from user can be used for that. As an example
is avocado vt which can use Cartesian configuration for resolving tests.
This commit adds the ability to discover tests from different data
without the test references.

Signed-off-by: Jan Richter <jarichte@redhat.com>