datadir.paths.data_dir now is defined from settings registration
The current code is creating a custom logic adding a default value that
is hardcoded. data_dir module was created way before our new settings
logic was introduced, default values now are defined at registration
time.  Also, with the current logic "avocado config --datadir" and
"avocado config" output are not consistent. On top of that, when users
define a custom value, in some cases Avocado is replacing this value
with a default hardcoded value, instead of respecting users' choice (for
instance when users have no permissions).  My proposal here is to use
`~/avocado/data` as the default data dir since usually, a data dir is
useful regardless of the method of installation. This is part of #4443,
and if approved I will work on the remaining datadir.paths.*. This is
following the same logic as 182e0c464f0e6872248718cd1655acb41652be9d.

Signed-off-by: Beraldo Leal <bleal@redhat.com>