Adjust requirement_package.py runner to reduce collision chances
On issue https://github.com/avocado-framework/avocado/issues/4620, when
two tasks try to fulfill a package, one is smart enough to wait for the
other to finish, but as soon as the lock is released, one may fail if the
other has still work to complete, like erase the cache.

This patch checks if the failure was not a false negative.

This does not fix #4620 as tests with the same requirement may still
collide. Easy to reproduce on s390x. But it reduces the chance of
happening on other architectures and allows some self-tests to be
enabled.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>