Adjust requirement_package.py runner to do not fail on collision
On issue https://github.com/avocado-framework/avocado/issues/4620, when
two tasks try to fulfill a package, one is smart enough to wait for the
other to finish, but as soon as the lock is released, one may fail if the
other has still work to complete, like erase the cache.

This patch checks if the failure was not a false negative.

Fixes: #4620

Signed-off-by: Willian Rampazzo <willianr@redhat.com>