avocado.core.job: Only store logger handler names
We don't need to store logger handler objects per
se in the protected Job.__logging_handlers dict,
we just need to store their names, since we can
easily retrieve the objects themselves by looking
in the logger.handlers attribute.

By doing that we can properly serialize the Job
object that is passed to multiprocessing, preventing
the error:

    avocado run /usr/bin/true
    JOB ID     : f419399f140000ddffa9e66095ce3e111895902e
    JOB LOG    : /Users/lmr/avocado/job-results/job-2021-06-06T10.48-f419399/job.log

    Avocado crashed: TypeError: cannot pickle '_thread.RLock' object
    ...

Running on MacOS and python3.9.

Signed-off-by: Lucas Meneghel Rodrigues <lookkas@gmail.com>