Sysinfo nrunner
The runner for gathering sysinfo. It supports the command and system
file collectible from the legacy runner and uses the
'sysinfo.collectibles' configuration. It is the part of work on #3877,
but it is just the runner and misses the configuration part in avocado
job.

You can test it by running the recipe in
'examples/nrunner/recipes/runnables/sysinfo.json'

Reference: #3877
Signed-off-by: Jan Richter <jarichte@redhat.com>