safeloader: remove legacy and unused find_class_and_methods() function
This function was once used for what it's named after: finding class
and method names in a Python source code file.

The problem with that function is that it is not being used to
actually power Avocado's safeloader method.  Worse, it was used in the
unittests, giving a false impression that all test classes in methods
on its own Python file were being properly found.

This removes this function, and some tests that depended on it.  Now
that the *actual* Avocado code to find tests is being used, and issue
to a relative import so that the "MultiLevel" class and its tests are
found, and the "FindClassAndMethods.test_self", which references all
tests in the file, passes.

Tests were removed here, but they covered unused code, so there should
be no loss in coverage.

Reference: https://github.com/avocado-framework/avocado/issues/4625
Signed-off-by: Cleber Rosa <crosa@redhat.com>