safeloader: avoid possibly using a variable before its definition
The _info, _disabled and _match variables may not have been defined in
the earlier block, and may always be set.

This changes the loop flow so that the "if _info" conditional will
only be reached with a _info variable properly defined.

This also makes _examine_class() more similar to find_python_tests()
behavior and style.

Signed-off-by: Cleber Rosa <crosa@redhat.com>