safeloader: support checking aliased statements on selftests
The "imported symbol" functionality does not care about the alias
names themselves.  But, it cares about properly parsing the "symbol"
and "module path" of all statements, including the ones that use
alias.

This splits the checks into a basic, that does not attempt to match
the (limited) statement created by to_str(), to the more complete
original one that contains the aliases.

Signed-off-by: Cleber Rosa <crosa@redhat.com>