safeloader: support multiple compatible statements in tests
Sometimes, it's not possible to tell, based on the "symbol" and
"module path" information, which exact statement may have generated
it.  For instance, both "from os import path" and "import os.path"
will effectively be about having the symbol "path", which should be
found at module path "os".

Let's allow multiple statements to be tested, and at the same time,
allow a match for a single the string representation.

Signed-off-by: Cleber Rosa <crosa@redhat.com>