safeloader: reimplement get_{symbol,module_path}_from_statement()
Both methods depend a lot on the same information that the other
method also uses.

For instance, to determine the final "module_path", we need to look if
there's a name to that will be a symbol.  So it makes sense to have
both things happening at the same place.

The old methods are kept for compatibility, but they can be dropped
soon.

Signed-off-by: Cleber Rosa <crosa@redhat.com>