safeloader: do not use the module directory in the import path list
One of the checks done when trying to find a suitable module (and
thus the interesting class on them) is if a name is found by the
Python import mechanism, and that uses a list of paths.

As it stands, we're manually adding the base directory of the
module we're investigating, which is not standard Python behavior.
That is, if you have:

  mod
  ├── a.py
  ├── b.py
  └── __init__.py

Code in "mod/b.py" can *not* do:

  >>> from a import symbol

And expect "symbol" do be found on "a".  For that, it should use
either:

 1) use relative imports, that is, "from .a import symbol"
 2) add "mod" to the sys.path/PYTHONPATH
 3) the current directory needs to be "mod"

This removes this non-standard behavior, and prevents modules being
found at non-expected places, and thus classes *not* being found on
those non-expected places.

Signed-off-by: Cleber Rosa <crosa@redhat.com>