Loader test files: move to a directory with an extra level
This introduces a new directory level, and renamed the
"loader_instrumented" to "safeloader".

The reason for the rename is that the code being tested with these
files also applies to Python unittests, so in general, it's really
about the safeloader.

The reason for the extra level of directory, is that it allows other
data files for other purposes to be put in the same directory.

Signed-off-by: Cleber Rosa <crosa@redhat.com>