safeloader: introduce a test for a caveat on multi level base class
There's currently a caveat/limitation in the safeloader code that
prevents a specific coding style when defining the base class
for an instrumented test: it can not came from a multi level
hierarchy.  This will currently not work:

  import project.base.module

  class MyTest(project.base.module.BaseTestClass):
  ...

While this limitation is easily resolvable by using arguably better
code styles like:

  from project.base import module

  class MyTest(module.BaseTestClass):
  ...

Or:

  from project.base.module import BaseTestClass

  class MyTest(BaseTestClass):
  ...

It's worth keeping the reproducer "documented" in a test.

Reference: https://github.com/avocado-framework/avocado/issues/4705
Signed-off-by: Cleber Rosa <crosa@redhat.com>