safeloader: change meaning of symbol
I believe I made a bad call with the original semantics of
the ImportedSymbol.  The original idea was to give an easy
access to the imported "symbol".  For instance, in an statement
such as:

   >>> import os

You only have one "thing" you're importing, so I initially thought
it was fair enough to call that the symbol being imported (it'd
end up being at the "symbol" attribute).

But, on "import" statements (as opposed to "from import" statements),
the only thing that can be imported is a module.

This changes the semantics, and makes it more clear what will
certainly be a module.

Signed-off-by: Cleber Rosa <crosa@redhat.com>