safeloader: use ImportedSymbol on to check for importability
This relies on the added support in ImportedSymbol to determine its
importability status, instead of using a private utility here.

Benefits include support for compound module paths.  Because of the
support for compound module paths, one of the selftests which would
not find Avocado tests, now finds Python unittests, whic is actually
also a bugfix.

It also removes all the compatibility methods with the previously
used "filesystem like" say for keeping track of the imported path,
module and symbol.

Signed-off-by: Cleber Rosa <crosa@redhat.com>