safeloader: follow imported symbols in addition to class definitions
While traversing the class hierarchy, we may find a given module
that does not contain a class definition (which would normally break
the path traversal), but that may contain an import for a symbol
matching the name we're looking for.

In that case, let's attempt to continue the traversal using that
imported symbol as a class.  From that point, it proceeds looking for
the matching module and class (say, "avocado" module and "Test"
class).

This fixes issue #4611.

Signed-off-by: Cleber Rosa <crosa@redhat.com>