Config variable for resolvers
The resolvers don't have access to the config variable, and they are
using `settings.as_dict()`. Unfortunately, this method returns only
config from command line and config files. When the configuration is
created by Job API, the `settings.as_dict()` doesn't respect that. This
leads to the incompatibility between resolvers config and suite config
created by `TestSuite.from_config()` in Job API.

This problem was hidden because avocado default resolvers don't need
config for resolving references and `TestSuite.from_config()` rewrites
the runnables config created by resolvers.

It is not such a big problem for avocado default resolvers, but it is a
problem for vt resolver which is using config in resolving process.

Signed-off-by: Jan Richter <jarichte@redhat.com>