Test: support for coroutines as tests
Python coroutines, AKA "async def" methods, are a common use case when
the code under test is based on asyncio.

Instead of forcing the user to come up with boiler plate code to run
each of the coroutines in a loop within a test, it's easier to
identify such methods and handle them accordingly.

One further possible improvement on top of this is to let the user
configure how the coroutine should be executed.  In this initial
implementation it simply gets one event loop and runs the coroutine
until it finishes.

Reference: https://github.com/avocado-framework/avocado/issues/4767
Signed-off-by: Cleber Rosa <crosa@redhat.com>