setup.py: remove conditional for absolute path
Here I argue that the check for the absolute nature of the plugin
directory is not a huge saving from the calculation of the absolute
path iself.

Getting the absolute path of an absolute path is almost a no-op,
with Python's abspath() including the check for an absolute path,
so let's simply use abspath().

Signed-off-by: Cleber Rosa <crosa@redhat.com>