setup.py: make action a required argument and influence action name
The walk_plugins_setup_py() function requires the action argument, so
let's reflect that on the function signature.  With that, we can infer
a better default to the the action name.

This also optimizes the setting of action names when they're are too
close to the default (CLEANING implicitly becomes CLEAN), and adds a
bit more context for the LINK/UNLINK (becomes DEVELOP LINK and DEVELOP
UNLINK).

Signed-off-by: Cleber Rosa <crosa@redhat.com>