Nrunner functional tests fix
Avocado has a problem with running two jobs which are connected to the
same status server. This problem is reported in #4308. We have two tests
which are affected by this issue because they are running their own
avocado jobs, and it is possible that those tests breaking Travis CI
which is  reported in #4768. This commit fixes those tests by creating
their own status server for their jobs.

Reference: #4308 #4768
Signed-off-by: Jan Richter <jarichte@redhat.com>