logging: preserving old root log level
We are messing up with the root log level. After finishing the run, we
need to restore back level. Because logging is a singleton, according to
the documentation we should be using a context manager for selective
logging, but unfortunately this will not be trivial today. So this should
fix it.

Signed-off-by: Beraldo Leal <bleal@redhat.com>