nrunner: add actual_time_start and actual_time_end
In some cases, such as the HTML plugin, we want to show the
actual time the test started or ended. If we use start_time
or end_time we'll get a monotonic time unrelated to the actual.
Here actual_time_start gets the data when nrunner starts
the test on StartMessageHandler and actual_time_end from
FinishMessageHandler when nrunner ends the tests.

Signed-off-by: Ana Guerrero Lopez <anguerre@redhat.com>