avocado.Test: prevent any test* method to be seen as a Python unittest
avocado.Test's teststmpdir property is badly named because it matches
what is expected to be a test.

Because Avocado inherits from Python unittest, and the extended
coverage for corner cases in the safeloader introduced, along with the
characteristic of the resolver that will give other plugins a chance,
this property is now being found as a Python unittest test method.

Besides this change, a method rename is desirable, but this will cause
API compatibility issues, which we guarantee until a new LTS release.
So, for now, let's disable all test* methods on avocado.Test from
being recognized as Python unittest methods.

Signed-off-by: Cleber Rosa <crosa@redhat.com>