Resolver and Discoverer plugins initialization fix
In the 43286b5fa3083fc0d3fb4381a4a40024d120eb81 the incompatibility with
Setting plugin was introduced. Because of that, avocado doesn't respect
plugin specific configuration. This commit adds `settings` variable
import to the `plugin_interfaces.py` Because of that, the `circular
import` bug is created when the plugins and Settings is initialized
during avocado initialization. This will make this import local to
specified plugins, which fix this problem.

Signed-off-by: Jan Richter <jarichte@redhat.com>