Software Manager: fix exraction of deb archive contents
The actual files that users will have access when installing deb
packages are not the contents of the "ar" archive, but the contents of
the "data.tar.*" tarball.

This fixes the extraction of the deb archives to actually extract the
data from the "data.tar.*" tarball.

The tests have skip conditions dropped because the implementation is
not dependant on the "ar" binary, but on the "avocado.utils.ar"
module.

Signed-off-by: Cleber Rosa <crosa@redhat.com>