nrunner: add support to failfast
Although nrunner has a distributed architecture and new concepts like
Workers and Spawners, we can deliver the failfast feature with some
limitations. Depending on the number of parallel tasks, some tasks might
have already started when an error occurs, because of that failfast will
work here in a "best-effort" assumption. This change will move all
non-started tasks to the finished queue when failfast is enabled.
Fixes #3870

Signed-off-by: Beraldo Leal <bleal@redhat.com>
Signed-off-by: Willian Rampazzo <willianr@redhat.com>