interfaces.py: Support get the interface details without IP version
Currently, if the interface does have not an IP address,
"_get_interface_details" will return an empty list, so this is a problem
to set default version=4. Most of the time, we only need the properties
of the interface itself, only "get_ipaddrs" will use IP addresses in
interface.py.

Signed-off-by: Yihuang Yu <yihyu@redhat.com>