Assets: support assigments with lists
One very important parameter of the fetch_assets() call is locations.
Because we currently only support the parsing of strings using the AST
parser, any example that uses locations fails to be recognized as a
asset for being fetched with `avocado assets fetch` or the equivalent
pre-tests plugin.

This adds support for list assignments, and simplifies the assets
example test so that it's easier to read, and supported under assets
fetch.

Fixes: https://github.com/avocado-framework/avocado/issues/4812
Signed-off-by: Cleber Rosa <crosa@redhat.com>