Test: make workdir initialization lazy
Avocado is currently creating the workdir directory at the class
initialization, which means that it's wasted for tests that get
skipped, canceled, or simply don't use them.

To avoid the extra overhead, and possibly leave stale dirs, let's make
the creation of workdir (and the ones used as the basis of workdir)
lazy.

For SIMPLE tests, the test itself can not request the creation of
workdir lazily, so it gets created manually.

Signed-off-by: Cleber Rosa <crosa@redhat.com>