Runner: set a command line parser for store logging stream
The `--store-logging-stream` parameter value is currently being
incorrectly parsed as a list of characters.  If a "bar" value is
given, it will generate the "b.INFO", "a.INFO", and "r.INFO" file.

This fixes the parsing of the command line arguments by treating the
value as a comma separated list, that becomes a set.

Signed-off-by: Cleber Rosa <crosa@redhat.com>