Settings: add support for different type for command line
Sometimes, for instance in the case of settings with lists, it's
not convenient at all to pass the same input expected by the config
file parser.  Example, on a config file you may have:

   [job.run]
   store_logging_stream = ['custom1', 'custom2']

But users would never be expected to come up with a command line
option such as:

   avocado run --store-logging-stream="['custom1', 'custom2']" ...

Instead, this is what users may expect to run:

   avocado run --store-logging-stream=custom1,custom2 ...

For this to possible, it's necessary to allow a different type for
an option, to be used exclusively for command line parsing.

Signed-off-by: Cleber Rosa <crosa@redhat.com>