nrunner: better validation with uri
While loadTestsFromName() accepts specifiers that can resolve either to a
module, a test case class or a test method within a test case class, for our
scenario, since a runner is responsible for running one single test, we need to
accept only uris with a proper test reference. This will abort in case a wrong
uri it is used.

Signed-off-by: Beraldo Leal <bleal@redhat.com>