nrunner: fixes loader.loadTestsFromName() resolution with absolute paths
Currently loader.loadTestsFromName() will only find tests that are part of
sys.path. We are adding current directory (.) to that path, assuming that test
references will always be a relative path to the current directory. With this
change we will be able to pass a absolute reference (i.e:
/tmp/foo.py:Bar.test_foo) and the nrunner will add the basedir to the sys.path
correctly. Fixes #4840, #4839

Signed-off-by: Beraldo Leal <bleal@redhat.com>