nrunner: add debugging aids to the Task/StateMachine/Repo system
The nrunner has a much more decoupled design with many more moving
parts (when compared to the legacy runner), for instance:

 * workers trying to get tasks to move over their life cycle stages
 * tasks being started by different spawner
 * tasks asynchronously reporting their status to status servers

It makes it non-trivial to debug where an Avocado internal failure may
be located.

This introduces the convention of the "avocado.core" logging
namespace, following the actual Python module implementation, so that
developers and advanced users can choose to see the nrunner internals
for troubleshooting, debugging, etc.

Because of the hierarchical nature of the namespaces, one can choose
to see all "avocado.core" information with:

  $ avocado --show=avocado.core run --test-runner=nrunner ...

Or choose a subset with:

  $ avocado --show=avocado.core.task run --test-runner=nrunner ...
  $ avocado --show=avocado.core.status run --test-runner=nrunner ...

By default, it stores the core, internal debugging information in a
separate file in the job results directory, because:

 * It's too much information that ends up polluting and disrupting
   "job.log"
 * It's really internal Avocado information, and not really meaningful
   to users expecting to see what happened to their tests in jobs
 * With the upcoming switch to nrunner being the default, having a
   log with internal information by default at least for the following
   few releases may help us to solve any issues users face with a
   readily available log file.

More (or less) debugging information can be added to the nrunner
related modules, or, for that matter, any other core module.

Fixes: https://github.com/avocado-framework/avocado/issues/4547
Signed-off-by: Cleber Rosa <crosa@redhat.com>