[exec runner] add support for the AVOCADO_TEST_* env variables
This adds support to the following environment variables for the
exec nrunner and those based on it (exec-test and tap):

AVOCADO_VERSION
AVOCADO_TEST_BASEDIR
AVOCADO_TEST_WORKDIR
AVOCADO_TEST_LOGDIR
AVOCADO_TEST_LOGFILE

This should be considered a compatibility patch to allow the same
behavior of the legacy runner. In this case, the LOGDIR and the LOGFILE
are created to have a pointer to them, inside the OUTPUTDIR folder. The
reason for that is because when running on a spawner that is not in the
host system, these paths will not match the host system's path.

Signed-off-by: Willian Rampazzo <willianr@redhat.com>