core/nrunner: add optional job_id to Tasks
This adds extra (but still optional) information to a Task, allowing
it to be tied to the job that has created it (if that's the case).

While a Task is currently only used by Avocado itself within the
context of a job, it's easier to enforce it being mandatory in the
future than it is to do the opposite.

Signed-off-by: Cleber Rosa <crosa@redhat.com>