core/nrunner: add variable for the default task category
The default category type for a task may be set from a number of
different places, including the Python API or the command line.  Also,
it is used in the message handling to check whether to include the
task results in the job results.

Let's define it so that it's reusable, and while at it, better
document it too.

Signed-off-by: Cleber Rosa <crosa@redhat.com>