status repo: require job ids and log mismatches
This binds a status repo to one job, logging mismatches when messages
are sent to the servers (and thus repos) that are intended for another
job.

This is intended to capture situations where users may manually create
Tasks (avocado.core.nrunner.Task) or even jobs (avocado.core.job.Job)
with configurations that result in clashes.

Signed-off-by: Cleber Rosa <crosa@redhat.com>