Podman spawner: default to using image matching the host OS
The primary goal of the Podman spawner is to provide more isolation to
tests, both for the sake of the tests, and for the protection of the
test execution environment (often a developer's workstation).

But the mismatch between host OS and the OS in the container are not
positive.  This will change the default so that the two of them match
as much as possible.

If it's impossible to detect the host OS, then the latest Fedora
version image will be used.

Signed-off-by: Cleber Rosa <crosa@redhat.com>