selftests.utils: reduce length (and detail) of temporary dirs
The very long and detailed temporary directory prefix used on
TestCaseTmpDir was introduced to quickly identify selftests that may
not be cleaning up after themselves.

But, it has one negative side effect: some files names such as sockets
have length limitations, and the amount of detail makes it very easy
to go beyond that limit.

This restricts the prefix name to the class name, which is usually a
good enough indication of the file containing the test.  The method
was pretty much useless, as it was always set to "setUp".

Signed-off-by: Cleber Rosa <crosa@redhat.com>