nrunner plugin: introduce status server auto mode
This introduces a mode under which the status server will set up its
own "listen" and "uri" values.

The "listen" value is of course related to where the status server
will be listening for status updates, from clients (tasks) which are
given the "uri" value.

Instead of a TCP based socket, the auto mode defaults to using a UNIX
domain socket created within the job's own directory, which better
avoids clashes (no two jobs should be using the same job directory
anyway).

One way to see the difference in action is to run:

 $ nc -l 127.0.0.1 8888

And on another session:

 $ avocado run --test-runner=nrunner --nrunner-status-server-auto /bin/true

And compare it to:

 $ avocado run --test-runner=nrunner /bin/true

But not using a resource available system wide, but one inside the
job's directory, the possibility of clashes are virtually
non-existing.

Signed-off-by: Cleber Rosa <crosa@redhat.com>