nrunner: enable automatic status server
Users often complain about jobs with no test results.  There are of
course many possible causes, but a common one is a clash and failure
of communication with the status server.

Because it's currently dependent on a fixed TCP port, clashes are
quite easy, especially when running multiple jobs at once (or nested).

Also, a good number of tests have to go through the trouble of setting
up custom status server to avoid clashes.  This removes those custom
and simply relies on the now automatic status servers.

This enables the automatic status server, based on a much more private
UNIX domain socket, so that clashes become virtually impossible.

Signed-off-by: Cleber Rosa <crosa@redhat.com>