Config for every runnable
The `Runnable` class has to have configuration because this information
are used by Tasks and Runners. When the runnable is created by
`from_config` method, everything works fine, but when the runnable is
created manually the config might be empty. This commit will set
runnables configuration same with the test suite configuration.

Reference: #4846
Signed-off-by: Jan Richter <jarichte@redhat.com>