CI/weekly.yml: duplicate the pipeline
The weekly pipeline using the python version in development
is failing. Looking at the log, it isn't whether clear
the failed tests are due to problems with the plugins or
something else.

To make easier to find where the problem is, this change
duplicates the pipeline in two different tests:
- first one, identical to the one run in every PR (with plugins)
- second one, without installing the optional plugins and
installing their test

Only the first job is run initially, if it fails, then the second
job is run.

When one of the job fails, the job-results log will be archived
as an artifact.

Export AVOCADO_LOG_DEBUG and AVOCADO_CHECK_LEVEL properly.
The section 'env' in the pipeline is used for job variables and
not environment variables.

Signed-off-by: Ana Guerrero Lopez <anguerre@redhat.com>