Job: enforce suite name uniqueness within the scope of a job
The test results uniqueness are guaranteed because of their unique
TestIDs, but the full result location of a test also takes into
account its suite name.

So far, there was no enforcement of unique suite names, allowing
results to be overwritten silently by tests of another suite with the
same name (if their TestIDs also match).

With this change, a very early and explicit exception will be raised,
allowing users of the Job API to fix their suite creation.

Signed-off-by: Cleber Rosa <crosa@redhat.com>