[utils] network/hosts: remove bonding_masters as a valid interface
If there is a bond interface on a machine, the bonding driver will create a file
in `/sys/class/net` called `bonding_masters`. When read, the file just lists all
the bonding interfaces present in the machine. i.e. `bond0 bond1`. Therefore,
`bonding_masters` should not be considered an interface.

Signed-off-by: Cris Forno <fornosoccer12@gmail.com>