asset: uses an unique identifier when name is not available
If an URL is providing multiple files (depending on the GET parameters), we
will replace the asset with the previous one downloaded because we are using
the same endpoint. For instance: `https://foo.bar/download?request=jasdf8d878`.
Fixes #4824.

For those cases, lets use an uuid4() as unique identifier of the asset.

Signed-off-by: Beraldo Leal <bleal@redhat.com>