fix env variable: s/AVOCADO_TEST_OUTPUT_DIR/AVOCADO_TEST_OUTPUTDIR/g
We do have an inconsistent variable name here. I could have chosen the reverse
exchange. But since I found more occurrences of the second name in our
documentation and in the examples, I decided to use this one. The only public
reference to the first one was on our man page in a case that I think is
safe-to-ignore.  Fixes #4901.

Signed-off-by: Beraldo Leal <bleal@redhat.com>