selftests/functional/test_output.py: test behavior of --show with nrunner
nrunner better isolates the output of tests it runs, even if those
tests are other Avocado jobs.  On the current state, the test expects
to find in the UI multiple jobs and tests that are "leaked" from the
other jobs (and the tests within them).

With the change to the nrunner, this adds a single and explicit
/bin/true execution to make sure that only the primary job output is
shown.

Signed-off-by: Cleber Rosa <crosa@redhat.com>