test_runner_timeout: check timeout from structured result output
Instead of checking the timeout condition from the xunit output thrown
on the standard output, let's use the structured JSON result file.

Because the output previously being used for the check was generated
by a component that only is only used for the legacy runner, the check
would not behave correctly under the nrunner.  While at it, also use
the complete test path as expected by the resolver.

Signed-off-by: Cleber Rosa <crosa@redhat.com>