selftests/functional/test_basic.py: remove loader specific check
These tests perform two different checks: one is specific to the
loader (that is, to the legacy runner), while the second is done at
the suite level, and thus applies to both the loader and resolver.

In preparation for the default runner switch, let's make the test work
for both runners.

Signed-off-by: Cleber Rosa <crosa@redhat.com>