prerelease.yml: update artifacts naming
Give different names to each job artifacts so they are all
kept without overwriting.

Do not save any artifact for the job check-rpm-available-copr,
since this job only install the packages and doesn't create
any output.

Update path for the artifacts generated by the job
avocado-deployment-copr, this job is running the tests
inside a container and the path changes.

Signed-off-by: Ana Guerrero Lopez <anguerre@redhat.com>