[PoC] Podman Spawner: use Avocado and setuptools eggs
Instead of using the "avocado/core/nrunner.py" script, this deploys
and uses Avocado and setuptools into the Podman container.

The avocado installation runs entirely from the eggs.

This is limited to containers that have Python 3.9, because these are
the only eggs currently built and saved (in a temporary location).

Reference: https://github.com/avocado-framework/avocado/issues/4934
Signed-off-by: Cleber Rosa <crosa@redhat.com>